requirements.txt
dotmap==1.3.30
imutils==0.5.4
tqdm==4.66.1 #Done
opencv-contrib-python==4.9.0.80
#Mask the face requirements
dlib==19.19.0
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.26.0
Pillow==12.1.0
torch==2.9.1
Requests==2.32.5
PyYAML==6.0.2
onnxruntime==1.23.2
onnxruntime-gpu==1.23.2
torchvision==0.24.1
scikit-image==0.25.2